Implement no-op event handling for native docker
Also, delete container state file when container is removed.
And refactor tests into multiple files.
