+++ tests/common_fixtures.py
@@ -7,15 +7,11 @@ from cattle import CONFIG_OVERRIDE
 CONFIG_OVERRIDE['DOCKER_REQUIRED'] = 'false'  # NOQA
 
 from os.path import dirname
-from datadiff.tools import assert_equals
 
 _file = os.path.abspath(__file__)  # NOQA
 sys.path.insert(0, dirname(dirname(_file)))  # NOQA
 
-from .response_holder import ResponseHolder
-from cattle import type_manager, plugins
-from cattle.agent import Agent
-from cattle.utils import JsonObject
+from cattle import plugins
 
 
 plugins.load()
@@ -33,55 +29,3 @@ os.makedirs(SCRATCH_DIR)
 def scratch_dir(request):
     request.addfinalizer(
         lambda: shutil.rmtree(SCRATCH_DIR, ignore_errors=True))
-
-
-@pytest.fixture(scope="module")
-def responses():
-    r = ResponseHolder()
-    type_manager.register_type(type_manager.PUBLISHER, r)
-    return r
-
-
-@pytest.fixture(scope="module")
-def agent(responses):
-    return Agent()
-
-
-def json_data(name):
-    marshaller = type_manager.get_type(type_manager.MARSHALLER)
-    with open(os.path.join(TEST_DIR, name)) as f:
-        return marshaller.from_string(f.read())
-
-
-def diff_dict(left, right):
-    for k in left.keys():
-        left_value = left.get(k)
-        right_value = right.get(k)
-        try:
-            diff_dict(dict(left_value), dict(right_value))
-            assert_equals(dict(left_value), dict(right_value))
-        except AssertionError, e:
-            raise e
-        except:
-            pass
-
-
-def event_test(agent, name, pre_func=None, post_func=None, no_diff=False):
-    req = json_data(name)
-    resp_valid = json_data(name + '_resp')
-
-    if pre_func is not None:
-        pre_func(req)
-
-    resp = agent.execute(req)
-    if post_func is not None:
-        post_func(req, resp)
-
-    if not no_diff:
-        del resp["id"]
-        del resp["time"]
-
-        diff_dict(JsonObject.unwrap(resp_valid), JsonObject.unwrap(resp))
-        assert_equals(JsonObject.unwrap(resp_valid), JsonObject.unwrap(resp))
-
-    return req, resp
