+++ tests/docker/native_container_activate
