+++ tests/docker/native_container_deactivate
