+++ tests/docker/native_container_volume_remove
@@ -1,0 +1,236 @@
+{
+  "id": "9e8fcc85-98dd-4d43-8f2f-9498cfbe906e",
+  "name": "storage.volume.remove",
+  "replyTo": "reply.3826413832580807098",
+  "resourceId": "1vspm6",
+  "resourceType": "volumeStoragePoolMap",
+  "publisher": null,
+  "transitioning": null,
+  "transitioningMessage": null,
+  "transitioningInternalMessage": null,
+  "previousIds": null,
+  "previousNames": null,
+  "data": {
+    "volumeStoragePoolMap": {
+      "name": null,
+      "id": 6,
+      "state": "removing",
+      "created": 1430504243000,
+      "description": null,
+      "kind": "volumeStoragePoolMap",
+      "uuid": "de425dde-f649-401b-b5bd-792b0a86b9ea",
+      "removed": null,
+      "volumeId": 6,
+      "removeTime": null,
+      "storagePoolId": 1,
+      "data": {
+        "fields": {}
+      },
+      "type": "volumeStoragePoolMap",
+      "volume": {
+        "name": null,
+        "id": 6,
+        "state": "purging",
+        "created": 1430504243000,
+        "uri": null,
+        "instanceId": 6,
+        "description": null,
+        "kind": "volume",
+        "accountId": 5,
+        "format": "docker",
+        "uuid": "913eca68-1ad0-4be3-bf11-b4946707ff8d",
+        "zoneId": 1,
+        "removed": 1430504265000,
+        "deviceNumber": 0,
+        "imageId": 6,
+        "offeringId": null,
+        "virtualSizeMb": null,
+        "physicalSizeMb": null,
+        "attachedState": "active",
+        "removeTime": 1430504325000,
+        "allocationState": "deactivating",
+        "data": {
+          "fields": {}
+        },
+        "type": "volume",
+        "offering": null,
+        "instance": {
+          "name": "prickly_darwin",
+          "id": 6,
+          "state": "purging",
+          "domain": null,
+          "created": 1430504243000,
+          "hostname": "f88709f01974",
+          "description": null,
+          "kind": "container",
+          "accountId": 5,
+          "compute": null,
+          "externalId": "f88709f019745276f904656443323c69b462cd2bba265a1c33e6b646354ef6bb",
+          "uuid": "03d997d5-f77b-43ae-9444-a880f0b498d4",
+          "zoneId": 1,
+          "instanceTriggeredStop": "stop",
+          "systemContainer": null,
+          "removed": 1430504265000,
+          "userdata": null,
+          "imageId": 6,
+          "memoryMb": null,
+          "offeringId": null,
+          "agentId": null,
+          "removeTime": 1430504325000,
+          "allocationState": "deactivating",
+          "firstRunning": 1430504244000,
+          "token": "4ZDlx5IB4ZMbbmMNRVrmiRPKdq3M3AFq6BAHUIDcjDq2uTnEw6qiziUZyj2V8t7bwdfyJrgAScbDA7wTA",
+          "registryCredentialId": null,
+          "nativeContainer": true,
+          "data": {
+            "fields": {
+              "requestedHostId": 1,
+              "imageUuid": "docker:ubuntu",
+              "networkIds": [
+                2
+              ],
+              "transitioningProgress": 95,
+              "dockerHostIp": "192.168.59.103",
+              "dockerPorts": {},
+              "dockerIp": null,
+              "domainName": null,
+              "user": null,
+              "memory": null,
+              "memorySwap": null,
+              "cpuShares": null,
+              "cpuSet": null,
+              "tty": true,
+              "stdinOpen": true,
+              "directory": null,
+              "environment": {
+                "RANCHER_NETWORK": "true"
+              },
+              "command": [
+                "bash"
+              ],
+              "entryPoint": [],
+              "dataVolumes": [],
+              "ports": [],
+              "privileged": false,
+              "publishAllPorts": false,
+              "lxcConf": {},
+              "dns": [],
+              "dnsSearch": [],
+              "capAdd": [],
+              "capDrop": [],
+              "devices": [],
+              "primaryIpAddress": "10.42.164.101"
+            },
+            "dockerInspect": null,
+            "dockerContainer": null
+          },
+          "type": "instance",
+          "dataVolumesFromContainers": []
+        },
+        "storagePools": [
+          {
+            "name": "boot2docker Storage Pool",
+            "id": 1,
+            "state": "active",
+            "created": 1430503112000,
+            "description": null,
+            "kind": "docker",
+            "accountId": 5,
+            "uuid": "220e8309-38c6-4158-b909-e98f9c5363f0",
+            "zoneId": 1,
+            "removed": null,
+            "agentId": 1,
+            "physicalTotalSizeMb": null,
+            "virtualTotalSizeMb": null,
+            "external": false,
+            "removeTime": null,
+            "data": {
+              "fields": {
+                "reportedUuid": "478d201e-f416-44a4-8a7e-169445e022d2-pool",
+                "hostUuid": "478d201e-f416-44a4-8a7e-169445e022d2",
+                "type": "storagePool"
+              }
+            },
+            "type": "storagePool"
+          }
+        ],
+        "image": {
+          "name": "ubuntu:latest",
+          "id": 6,
+          "state": "active",
+          "created": 1430504243000,
+          "description": null,
+          "url": null,
+          "kind": "image",
+          "accountId": null,
+          "format": "docker",
+          "uuid": "36509ba1-7ec1-4e78-a18d-63fefbb872f2",
+          "instanceKind": "container",
+          "removed": null,
+          "isPublic": false,
+          "virtualSizeMb": null,
+          "physicalSizeMb": null,
+          "checksum": null,
+          "removeTime": null,
+          "registryCredentialId": null,
+          "data": {
+            "dockerImage": {
+              "id": null,
+              "repository": "ubuntu",
+              "namespace": null,
+              "tag": "latest",
+              "server": "index.docker.io",
+              "fullName": "ubuntu:latest",
+              "qualifiedName": "ubuntu",
+              "lookUpName": "ubuntu"
+            },
+            "fields": {}
+          },
+          "type": "image"
+        }
+      },
+      "storagePool": {
+        "name": "boot2docker Storage Pool",
+        "id": 1,
+        "state": "active",
+        "created": 1430503112000,
+        "description": null,
+        "kind": "docker",
+        "accountId": 5,
+        "uuid": "220e8309-38c6-4158-b909-e98f9c5363f0",
+        "zoneId": 1,
+        "removed": null,
+        "agentId": 1,
+        "physicalTotalSizeMb": null,
+        "virtualTotalSizeMb": null,
+        "external": false,
+        "removeTime": null,
+        "data": {
+          "fields": {
+            "reportedUuid": "478d201e-f416-44a4-8a7e-169445e022d2-pool",
+            "hostUuid": "478d201e-f416-44a4-8a7e-169445e022d2",
+            "type": "storagePool"
+          }
+        },
+        "type": "storagePool"
+      }
+    }
+  },
+  "time": 1430504369732,
+  "timeoutMillis": 15000,
+  "transitioningProgress": null,
+  "context": {
+    "processId": "121",
+    "logicName": "VolumeStoragePoolMapRemove",
+    "logicPath": "volume.purge->(VolumePurge)->volume.deallocate->(VolumeDeallocate)->volumestoragepoolmap.remove->(VolumeStoragePoolMapRemove)",
+    "processName": "volumestoragepoolmap.remove",
+    "prettyResource": "volume:6->volumeStoragePoolMap:6",
+    "topResourceType": "volume",
+    "resouceId": "6",
+    "resouceType": "volumeStoragePoolMap",
+    "topProcessName": "volume.purge",
+    "prettyProcess": "volume.purge->volumestoragepoolmap.remove",
+    "topResourceId": "6",
+    "processUuid": "31760690-79e8-454e-a55b-8c2bd4bcf7ee"
+  }
+}
